put related security group filter directly on the VPC resource
